Several reliability fixes a full audit of the new_sat_tracker.cc to bring the code up to spec fixed a build error in Vallado's SGP4 code a re-write of the Kindex parser to accomodate additional changes per NOAA Service Change Notice 26-21 this also involves a data source change https://services.swpc.noaa.gov/products/solar-wind/plasma-7-day.json to https://services.swpc.noaa.gov/json/rtsw/rtsw_wind_1m.json

+++ utils/conversions.cc
@@ -39,7 +39,7 @@ bool isLeapYear(int year) {
     return (year % 400 == 0) || (year % 4 == 0 && year % 100 != 0);
 }
 
-void doy_to_mmdd(const int year, int doy, int* mm, int* dd) {
+void doy_to_mmdd(const int year, uint16_t doy, int* mm, int* dd) {
     static int months[14]={0, 31, 28,  31, 30, 31, 30, 31, 31, 30, 31, 30, 31, 0};
     if ((!mm) || (!dd)) {
 //        std::cout << "Invalid result locations\n";
